+++ src/main/java/com/invadermonky/magicultureintegrations/integrations/roots/RootsAgricraft.java
@@ -1,5 +1,0 @@
-package com.invadermonky.magicultureintegrations.integrations.roots;
-
-public class RootsAgricraft {
-    //TODO: Roots elemental soil working with agricraft crops
-}